Dev/products/Idea_11_Code_Review_Bias_Detector/requirements.txt
fastapi
uvicorn
httpx
python-multipart
sqlalchemy
psycopg2-binary
pandas
numpy
python-dotenv
anthropic
pydantic
python-jose[cryptography]
